requirements.txt
Django==2.0.8
dj-database-url==0.5.0
psycopg2==2.7.5
wagtail==2.2.2
django-heroku
django-cors-headers
django-select2
django-admin-select2
python-slugify